chore: split plugin code into several files
