+++ src/pupil_labs/neon_player/plugins/events/__init__.py
@@ -1,0 +1,4 @@
+from .plugin import EventsPlugin, EventsPluginGlobalProps
+
+
+__all__ = ["EventsPlugin", "EventsPluginGlobalProps"]
